create forcing workflow
