+++ hydromt_delwaq/workflows/__init__.py
@@ -4,3 +4,5 @@
 from .emissions import *
 from .segments import *
 from .roads import *
+from .hydrology import *
+from .forcing import *
